Convert display-lazy-ads to TypeScript
